--- lib/iris/__main__.py
@@ -1,12 +1,0 @@
-# Copyright Iris contributors
-#
-# This file is part of Iris and is released under the LGPL license.
-# See COPYING and COPYING.LESSER in the root of the repository for full
-# licensing details.
-
-"""Iris module entry point."""
-
-from .experimental.ugrid import main
-
-
-main()
