+++ requirements/ci/py38.yml
@@ -16,6 +16,7 @@ dependencies:
   - cftime<1.3.0
   - click
   - click-default-group
+  - colorama
   - dask>=2
   - matplotlib=3.3.4
   - netcdf4
