+++ requirements/core.txt
@@ -5,6 +5,7 @@ cf-units>=2
 cftime<1.3.0
 click
 click-default-group
+colorama
 dask[array]>=2
 matplotlib
 netcdf4
